AMQP-505: Add setUri() method to ConnectionFactory
Previously the setUri() method on rabbitmq.ConnectionFactory was not exposed
to the CachingConnectionFactory nor the SingleConnectionFactory. This change
adds this method to the AbstractConnectionFactory, which is then inherited by
the two implementation classes.

A constructor taking the URI is added to CachingConnectionFactory and also to
the SingleConnectionFactory to short-circuit the construction even further.

AMQP-505: Polishing