AMQP-518: Remove <listener-container/> id Attrib.
JIRA: https://jira.spring.io/browse/AMQP-518

Previously, we deprecated the ID (AMQP-416) but simply ignoring it
is not really a good solution because the user "discovers" the bean
name changes (e.g. when autowiring).

Remove the attribute instead to force a fast failure.

Leave the attribute in the schema as a comment, in case users look there
first after seeing the problem after an upgrade.

Beef up the what's new to an `IMPORTANT` block.

AMQP-518: Remove <listener-container/> id Attrib.

JIRA: https://jira.spring.io/browse/AMQP-518

Previously, we deprecated the ID (AMQP-416) but simply ignoring it
is not really a good solution because the user "discovers" the bean
name changes (e.g. when autowiring).

Remove the attribute instead to force a fast failure.

Leave the attribute in the schema as a comment, in case users look there
first after seeing the problem after an upgrade.

Beef up the what's new to an `IMPORTANT` block.

AMQP-518: Fix Tests