Fix x-expires Test Case
Due to the small `x-expires`, it's possible for both consumers to
redeclare the queue; the test only expects one to do so.

Also, change the queue to have a random name to avoid collisions with other builds.

Also, noticed an NPE occasionally when running the test (null `BackOffExecution`).

Polishing

Fix the test instead of avoiding the NPE which can't happen in real code.