AMQP-546: Add SmartLifecycle to Caching Conn Fact
JIRA: https://jira.spring.io/browse/AMQP-546

Code was added to prevent connectoon use after the context is closed. However,
this was too early in the lifecycle because other beans might not have been
stopped yet; they could no longer use the connection.

Add `SmartLifecyle` to `CachingConnectionFactory` and put it in a late phase,
by default, so that beans in earlier phases can still use the connection in
`stop()`.

Add test case to reproduce the problem.