AMQP-535: PublisherCallbackChannel Improvements
JIRA: https://jira.spring.io/browse/AMQP-535

Internal state from the `PublisherCallbackChannel` (pending confirms) leaked
into the `RabbitTemplate` for the sole purpose of supporting `getUnconfirmed()`
which expires old entries and returns them to the caller.

Encapsulate this state within the channel and do not expose externally.

- Change `addListener` to return `void`
- Delete the unused `removeListener()` method from `PublisherCallbackChannel`
- Add `expire()` to `PublisherCallbackChannel`
- Change `getUnconfirmed` to invoke `expire` on each channel
- Change `PublisherCallbackChannel.Listener.removePendingConfirmsReference` to `...revoke()`
  (used to tell the Listener that it will receive no more confirms or returns from this channel
- Fix a problem in the `CachingConnectionFactory` where the pending confirms are not nacked
  if the underlying channel was detected as closed in `logicalClose()`

Polishing - PR Comments

Polishing - Fixes

Fix putIfAbsent in addListener to test for is absent (was incorrectly not absent)

The previous commit removed synchronization on the pendingReplies value.
This was required because the value was a synchronized map, the contract of which
requires synchronization while iterating.

Change the map value to a `ConcurrentSkipMap` which requires no synchronization.

* Fix inconsistent Mockito usage in the `LocalizedQueueConnectionFactoryTests`: https://build.spring.io/browse/AMQP-MEIGHT-595