Polishing - Fixes
Fix putIfAbsent in addListener to test for is absent (was incorrectly not absent)

The previous commit removed synchronization on the pendingReplies value.
This was required because the value was a synchronized map, the contract of which
requires synchronization while iterating.

Change the map value to a `ConcurrentSkipMap` which requires no synchronization.