AMQP-549: @RabbitListener Inferred Conversion Type
JIRA: https://jira.spring.io/browse/AMQP-549

When receiving messages with JSON content from foreign systems (without type
information in headers), the `@RabbitListener` is not convenient.

With this change, the type of the argument is inferred (if possible) from the
target method signature.

Methods with a single unannotated parameter or one with a `@Payload` annotation
are considered for this feature.

If the type cannot be inferred, a customized message converter can use the `Method`
object to determine the type itself.

This feature is not supported by `@RabbitHandler` methods (class-level `@RabbitListener`)
because the method selection is performed after the payload has been converted.

In addition to the inferred type and method, the target bean is also made available.
This might allow a custom converter to determine the type, even for a class-level
annotation.

In summary, when using a `@RabbitListener`, 3 new properties are added to `MessageProperties`:

- `inferredArgumentType` (if it can be determined)
- `targetMethod` (for method-level annotations)
- `targetObject` (in all cases)

Polishing - PR Comments

Also Skip Channel, Message<?> Parameters

Provided args are Message<?>, Message, Channel.

Since these are post-conversion arguments, they should not be
considered when inferring the argument type.

Polishing - More PR Comments

Speed up Test Shut Down

Polishing

Support Generic Spring Messaging Message