AMQP-591: Channel Cache (with Permits) Backport
JIRA: https://jira.spring.io/browse/AMQP-591

Permits are not released after a connection loss.

Cached channels held a reference to the old connection proxy; since this
was the key to the checkoutPermits map, the map entry was not found
when closing, and hence the permits were not increased.

The solution is to change the channel proxy to be permanent, and just refresh
its target connection; remove the hashCode method so cached channels can always
find their permits. Previously the hashcode included the hashcode of the delegate.

For `CacheMode.CONNECTION`, when a new proxy is created it is permanently stored
in `allocatedConnections` (previously called `openConnections`). When a connection
is retrieved from the cache, if the underlying physical connection is closed we
look at the next cached connection. If all cached connections are closed, we
attempt to refresh the underlying connection.

Also, the permits map entry for a connection was set up after the
listeners were called, so the admin, for example didn't consult the permits
at all.

Added an illegal state exception if the permits are not found when checking out as
well as error logging during return.

Add a test case.

Polishing

We now never remove proxy classes; when the connection cache limit is exceeded,
we leave the allocated proxy and keep it in the idle cache. We do, however, ensure
that the over-the-limit connection is closed, but the proxy will be reused later,
thus retaining the same map key value.

Finally, added a new property `connectionLimit` which limits the total number
of connections allowed.

Also, reverted the cache stats to show the actual number of open connections.

Fix Test Comment

Conflicts:
	spring-rabbit/src/main/java/org/springframework/amqp/rabbit/connection/CachingConnectionFactory.java
	spring-rabbit/src/test/java/org/springframework/amqp/rabbit/connection/CachePropertiesTests.java
	src/reference/asciidoc/amqp.adoc
	src/reference/asciidoc/whats-new.adoc

Resolved.