Polishing
We now never remove proxy classes; when the connection cache limit is exceeded,
we leave the allocated proxy and keep it in the idle cache. We do, however, ensure
that the over-the-limit connection is closed, but the proxy will be reused later,
thus retaining the same map key value.

Finally, added a new property `connectionLimit` which limits the total number
of connections allowed.

Also, reverted the cache stats to show the actual number of open connections.