AMQP-572: Always commit only in `afterCompletion`
JIRA: https://jira.spring.io/browse/AMQP-572

Since RabbitMQ doesn't support XA transactions and it really doesn't care about the transaction boundaries,
it doesn't matter when we call `channel.basicAck()` and `channel.txCommit()`.
Therefore any `TransactionManager` looks the same for RabbitMQ and we don't need to distinguish if we are within a local transaction or XA.

* Move `this.resourceHolder.commitAll()` in the `RabbitResourceSynchronization` to the `afterCompletion()` permanently, unconditionally.
* All test passes and now the ACK on the RabbitMQ will be done only if the commit to the DB is successful, for example.
* And it doesn't matter now if we use  XA transactions or Dave Syer's `Best Efforts 1PC pattern` with the `ChainedTransactionManager`

NOTE: There might be some breaking effect for end application if they do some "hard" work around TX synchronizations and rely on the previous behavior.
Since it is minor release, I think we can live with that and provide some workaround on demand.