RabbitTemplateIntegrationTests: fix timing issue
https://build.spring.io/browse/AMQP-MEIGHT-711

The `RabbitTemplateIntegrationTests.testSymmetricalReceiveAndReply()` use `this.template.receive()` which is based on the `receiveTimeout`.
Since `receiveTimeout == 0` by default that means "receive-no-wait".
and there is no guaranty that the message appears in the queue exactly after the previous `receiveAndReply()`.
The existing `this.template.setReplyTimeout(10000);`in the test does not make effect for the `receive()` operations.

* Specify `this.template.setReceiveTimeout(10000);` to meet test requirements and have some time window to wait for the message in the queue.