AMQP-624: Add SSL Support to Logging Appenders
JIRA: https://jira.spring.io/browse/AMQP-624

Adds the ability to configure the log4j2 and Logback
AMQP appenders to use SSL. As it uses a RabbitConnectionFactoryBean
to configure create the ConnectionFactory, the same options
when configuring SSL for the factory bean are available and
the same default values are used for the SSL configuration.

Fixes log4j2 Appender. Add SSL documentation

- Check boolean result for manager.activateOptions()
- Add SSL information to logging documentation
- Removed all deprecated log4j info.

* Fix overflow for the long `clientConnectionProperties` literal in the `logging.adoc`