AMQP-644: Non-Fatal Exception Stopped Container
JIRA: https://jira.spring.io/browse/AMQP-644

Previously, any `NoSuchMethodException` was container-fatal (the container was stopped)
on the assumption that the listener was mis-configured.

However, this could be a transient, data-related, problem so should be considered
message-fatal, not container-fatal.

Also added `ClassCastException` because when the `MessageListenerAdapter` delegate is a
lambda, we get this exception instead of a `NoSuchMethodException`.

* Remove lambda test because language level isn't 1.8 yet