AMQP-648: Fix Race in handleCancel (SMLC)
JIRA: https://jira.spring.io/browse/AMQP-648

The `cancelled` boolean is set when there are no more consumers in `handleCancelOk`.

Consider a case where a container is configured for multiple queues; one queue is deleted
from the broker. In `handleCancel`, we attempt to cancel the rest of the consumers so the
container can recover.

However, if those cancels fail (e.g. if the broker goes down), we'll never finalize the
consumer.

Move the consumer removal (and setting `cancelled`) to `basicCancel` instead. We don't really
care whether or not we get the `cancelOk`s.

Add mock test - also tested with the SI sample modified for 2 queues
- set a breakpoint in `handleCancel`
- delete a queue
- stop/start rabbitmq
- release the breakpoint
- container recovers

(cherry picked from commit 2328ee2)