TemplPubCITests2: Fix static field race condition
https://build.spring.io/browse/AMQP-SA16X-19/

The current `PublisherCallbackChannelImpl` logic is based on the logic around `static` fields for RabbitMQ Client 3.6.
Those fields for methods are populated from the provided `Channel`instance.
Since there are a lot of tests with mocks for target `Channel`, there is a case when the `RabbitTemplatePublisherCallbacksIntegrationTests2` fails with the `object is not an instance of declaring class`, because the real AMQP `Channel` isn't an instance of previous mock.

Add `@BeforeClass` to the `RabbitTemplatePublisherCallbacksIntegrationTests2` to reset `conditionalMethodsChecked` static field before test-case to have other fields for methods to be refreshed on the next invocation during this test