AMQP-655: RabbitTemplate and DirectReply Container
JIRA: https://jira.spring.io/browse/AMQP-655

Use a `DirectReplyToMessageListenerContainer` for Direct reply-to by default.

Add an option to revert to previous behavior of creating a new consumer for each
request.

Fix a problem in the `AmqpInvokerServiceExporter` where the correlation id was not
returned.

Fix a test `testSendToMissingExchange` - sometimes we can get a connectio already closed
exception.