AMQP-616: Direct ReplyTo Container Race Condition
JIRA: https://jira.spring.io/browse/AMQP-616

Consider the case where a reply is received at the same time as the future times
out or is canceled. At the same time, another request gets the same channel.

It is possible that the "late" future timeout/cancel will remove the consumer from
the in-use map even though the "next" user is waiting for the reply.

Add an epoch to the consumer which is incremented on each use. When releasing
the consumer, only do so if the epoch matches.