AMQP-687: Fix Conditional Rollback
JIRA: https://jira.spring.io/browse/AMQP-687

When using a rule-based transaction attribute with a rollback/no-rollback rule,
the rule is ignored when providing the transaction attribute directly to the container
rather than via a transaction advice,

Also, user exceptions are always wrapped in a `ListenerExecutionFailedException` which
limits the ability to configure rules, even when using an advice.

- Honor the attribute `rollBackOn()` mechanism within the container
- Add `ListenerFailedRuleBasedTransactionAttribute` to allow the user to configure rules based on the actual exception.

Add docs.

AMQP-686: Handle ImmediateAcknowledgeAmqpException

JIRA: https://jira.spring.io/browse/AMQP-686

Previously, `ImmediateAcknowledgeAmqpException` only worked if thrown by
an error handler, not a listener.

Examine the cause chain to see if this exception appears, and handle accordingly.

An `AmqpRejectAndDontRequeueException` found first stops the chain traversal.

Also clarify the javadocs to make it clear the current transaction is ended
immediately, not all message delivery.

checkstyle

Polishing and Fixes

While polishing for PR comments, I found some more problems (existing code).

When using an external transaction manager, txRollback() was called twice, once
within the transaction and once by the after completion in the transaction synchronization.

This was benign because the delivery tags were not registered, but incorrect.

Now, the delivery tag is always registered and the rollback done by the synchronization.

Added more tests and added tests for the local transaction scenario.

Add missing .this

Polishing