Polishing and Fixes
While polishing for PR comments, I found some more problems (existing code).

When using an external transaction manager, txRollback() was called twice, once
within the transaction and once by the after completion in the transaction synchronization.

This was benign because the delivery tags were not registered, but incorrect.

Now, the delivery tag is always registered and the rollback done by the synchronization.

Added more tests and added tests for the local transaction scenario.