DirectReplyToMLContainerTests race condition
https://build.spring.io/browse/AMQP-MASTER-645/

The logic in the test is based on the fact that after receiving a reply message we free the `Channel` and a new `container.getChannelHolder()` will return the same instance.
But since we perform `latch.countDown()` directly in the `MessageListener`, a desired `inUseConsumerChannels.remove(channel)` is called a bit later in the `final` of outer `ChannelAwareMessageListener`.

* Extract wrapper `ChannelAwareMessageListener` from the container and wrap it into one more listener to perform `latch.countDown()` exactly after a desired `inUseConsumerChannels.remove(channel)`

NOTE: Mockito doesn't work on final, anonymous classes or primitive types.