AMQP-711: Make Transaction Rollback Consistent
JIRA: https://jira.spring.io/browse/AMQP-711

Previously, messsages were unconditionally requeued when an external
transaction manager is used; this was not consistent with local
transactions, where the normal requeue logic was honored.

Fix the case when the transaction manager is a RabbitTransactionManager

When the transaction manager is a `RabbitTransactionManager`, the resource is already bound and
the resource bound within the transaction template execute() method is ignored.

The `requeueOnRollback` field needs to be set in the correct resource holder.

Change the `ConnectionFactoryUtils.bindResourceToTransaction` method to return the actual
resource holder that is bound to the transaction and set the boolean therein.

Polishing - clear the transaction thread locals.

Handle corner case where we have a transaction manager, but no global transaction.

This can only happen if the transaction attribute causes a transaction to be not started.

In this case, act like a local transaction, wrt committing and rolling back.

Test Polishing

The verify on commit/rollback should not be conditional.

Conflicts:
	spring-rabbit/src/main/java/org/springframework/amqp/rabbit/listener/AbstractMessageListenerContainer.java
	spring-rabbit/src/main/java/org/springframework/amqp/rabbit/listener/DirectMessageListenerContainer.java
	spring-rabbit/src/main/java/org/springframework/amqp/rabbit/listener/SimpleMessageListenerContainer.java
	spring-rabbit/src/test/java/org/springframework/amqp/rabbit/listener/ExternalTxManagerTests.java
	src/reference/asciidoc/whats-new.adoc
Resolved.