Fix the case when the transaction manager is a RabbitTransactionManager
When the transaction manager is a `RabbitTransactionManager`, the resource is already bound and
the resource bound within the transaction template execute() method is ignored.

The `requeueOnRollback` field needs to be set in the correct resource holder.

Change the `ConnectionFactoryUtils.bindResourceToTransaction` method to return the actual
resource holder that is bound to the transaction and set the boolean therein.