AMQP-717: RabbitTemplate Invoke Method
JIRA: https://jira.spring.io/browse/AMQP-717

Provide a mechanism to run `RabbitTemplate` opertions in a dedicated channel
instead of returning the channel to the cache on each operation.

Facilitates low-level channel operations like `waitForConfirmsOrDie()`.

For this reason, we need to subvert the channel proxy refresh logic and use the
bare channel in this mode.