AMQP-749: Force Close Channel With Slow Consumer
JIRA: https://jira.spring.io/browse/AMQP-749

If the consumer thread does not respond to a shutdown (e.g. "stuck" in user code),
the channel remains open and messages remain in an unack'd state.

Add a boolean (default true in 2.0) controlling the behaviour and allowing the container
to force close the channel - thus requeuing any unack'd messages.

Since this code runs on the thread that calls `stop()`, we need to always be sure that
the physical close required `ThreadLocal` is cleared.

There are test cases where a mix of mock and real channels are used and the TL was dirty, causing
side-effects; but this could really be a concern for a real application too.

Polishing - PR Comments

Fix Tests

Conflicts:
	spring-rabbit/src/main/java/org/springframework/amqp/rabbit/core/RabbitTemplate.java
	spring-rabbit/src/main/java/org/springframework/amqp/rabbit/listener/DirectMessageListenerContainer.java
	spring-rabbit/src/main/java/org/springframework/amqp/rabbit/listener/SimpleMessageListenerContainer.java
	src/reference/asciidoc/whats-new.adoc
Resolved.