GH-632 Add RepublishMessageRecoverer.deliveryMode
Fixes spring-projects/spring-amqp#632

By default all messages is sent with the `PERSISTENT` delivery mode,
unless we specify it manually otherwise.
We set it to `null` on the received message.
When it is `null` it is treated on the Broker as `NON_PERSISTENT`.
That is unexpected behavior, especially when received message is with
the `PERSISTENT` delivery mode.

* Add `RepublishMessageRecoverer.setDeliveryMode()` to specify a
default value
* Add getter for properties and make `final` properties as `protected`
to allow access from the inheritors
* Document new `setDeliveryMode()` and how properties can be overridden
from the `additionalHeaders()` extension