GH-672: Fix SpEL parsing for @RabbitListener
Fixes spring-projects/spring-amqp#672

The `StandardBeanExpressionResolver` is based on the
`TemplateAwareExpressionParser` which can parse any string to the
`LiteralExpression` or to the `CompositeStringExpression` if
template tokens (`#{` & `}`) are present in the source expression.
This way we don't need to deal with literal concatenations within
templated expression.
The expression like `#{'foo.' + myBean.bar}` can be replaced
with the `foo.#{myBean.bar}`