AMQP-833: DLQ and Expiry - Discard
JIRA: https://jira.spring.io/browse/AMQP-833

A common pattern is to add time-to-live to a DLQ and route
expired messages back to the main queue.

If a message fails with a fatal exception, such messages will
cycle around for ever; the application has no mechanism to
look at the `x-death` header.

Add logic to the `ConditionalRejectingErrorHandler` to look for
an `x-death` header and discard the message completely by throwing
an `ImmediateAcknowledgeAmqpException`.

Add a property to the error handler to revert to the previous
behavior.

* Polishing Docs according PR comments