AMQP-848: RT with DRTMLC - always release consumer
JIRA: https://jira.spring.io/browse/AMQP-848

When using a `DirectReplyToMessageListenerContainer` the consumer was released
for reuse after a normal `onMessage` or if an exception is thrown.

Neither of these occur when the reply times out.

Move the release to a `finally` block - it is idempotent so won't affect the
normal `onMessage` case.

**cherry-pick to 2.0.x**