Modernizing the look and feel and removing Docbook
I set things up to match the new standard we have for producing reference guides. Specifically, it now fetches the necessary resources from spring-doc-resourcesand uses them to put a collapsing ToC on the left and use the modern look and feel. I also drove out Docbook and an old stylesheet.