Formatting changes suggested by Artem
Artem asked for "Authors: " to appear before the author list. He also wanted to drop "Version" and bold the version number. Those were all easy changes. He would like to move the legal page in the PDF output (and I agree), but the PDF generator has a bug there: That page can't be moved without modifying the PDF with a third-party tool or tinkering with Asciidctor's output.