GH-923: Fix start delay for @Lazy @RabbitListener
Fixes https://github.com/spring-projects/spring-amqp/issues/923

While checking for missing or mis-matched queues, a lazily-loaded
listener container can deadlock for 60 seconds.

This occurs if the `allBeanNamesByType` cache does not currently have
an entry for `Queue` (e.g. cleared by registering a singleton).
When lazy beans are referenced, the `RabbitListenerEndpointRegistry`
starts the container and `start()` waits for the consumers to start.
Getting a reference to the lazy bean holds the `singletonObjects`
lock, which is required by the consumer(s) to get the `Queue` beans
to check.

Add a test case to demonstrate the issue.

Disable the redeclaration logic during the initial start of such a
container.

**cherry-pick to 2.1.x**