Add managed dependency versions to generated POM
Since the dependendency management Gradle plugin is applied and the
Jackson BOM is used, versions for this artifact are resolved by the
plugin.

The plugin configuration in this build disables POM customization, as
otherwise it would add `<dependendencyManagement>` sections in the
generated POMs.

This commit adds a missing script in the 2.1.x branch that checks the
generated POM for missing versions in dependencies and uses the
information provided by the dependency management plugin to add those
back.

Fixes gh-1082