GH-1093: @RabbitListener fix abstract return type
Fixes https://github.com/spring-projects/spring-amqp/issues/1093

AMQP-807 added support for generic return types; however this broke abstract
return types since the `__TypeId__` header was set to the abstract type.

If the return type is not a container type and represents an abstract class
or interterface use the concrete return type to construct the `JavaType`.

**cherry-pick to 2.1.x**