GH-1136: Support Meta Annotations with AliasFor
Resolves https://github.com/spring-projects/spring-amqp/issues/1136

Note that this is a minimal implementation that avoids major refactoring
of the BPP.

Repeatable user meta-annotatiions are not supported.

In a future release we should perform that refactoring as well as supporting
`@Repeatable` user annotations.

Also restore the `stop()` with finally to fix `EnableRabbitTests.testProperShutdownOnException()`
which took 30 seconds to run after the finally was removed by
c1e31798961262795c4aa2b4bab3f0f67fedda26