GH-1201: Native BatchMessageListener Support
Resolves https://github.com/spring-projects/spring-amqp/issues/1201

Support de-batching producer batches to a `List<Meessage>` with native
`BatchMessageListener`s; previously this was only possible with
`@RabbitListener` which does the debatching itself.

* Fix race in test.

https://github.com/spring-projects/spring-amqp/pull/1202#issuecomment-629436933

The test has a very short timeout; there is a race such that the channel won't be
physically closed if we don't get the `ConsumeOK` in time.

This won't really cause a problem in a real application, but this change will
prevent the test from sporadically failing.
# Conflicts:
#	src/reference/asciidoc/amqp.adoc