GH-1230: Fix return/confirm delivery order
Resolves https://github.com/spring-projects/spring-amqp/issues/1230

We must guarantee that returns are delivered before confirms.
https://github.com/spring-projects/spring-amqp/issues/1055 changed the logic
to deliver on a different thread to avoid a deadlock in the amqp-client.

Since the executor service might have multiple threads, we added a latch to
ensure the return is delivered first.

Unfortunately, this was a one-shot latch and subsequent deliveries using the
same channel might arrive in the wrong order.

Move the latch to the `PendingConfirm` to reinstate in-order delivery.

**cherry-pick to 2.2.x**